requirements.txt
djangorestframework~=3.11.0
Django~=3.1
djoser~=2.0.3
django-admin-sortable2~=0.7.6
django-filter~=2.3.0
pinyin~=0.4.0
phonenumbers~=8.12.7
geopy~=2.0.0
django-admin-sortable~=2.2.3
python-dotenv~=0.14.0
django-phonenumber-field~=4.0.0
djangorestframework-simplejwt~=4.4.0
Pillow~=7.2.0
drf-yasg==1.17.1
